feat: adiciona componente TituloConfig - para titulos da área de configuração de perfil
