+++ src/pages/app/perfil/dados.tsx
@@ -1,25 +1,12 @@
 import React from 'react'
-import { Box, Container, makeStyles, Typography } from '@material-ui/core'
+import { Container } from '@material-ui/core'
 import { withUser } from 'src/components/hocs/withAuth'
-import LinkVoltar from 'src/components/LinkVoltar'
 import NavigationList from 'src/components/NavigationList'
 import { useRouter } from 'next/router'
 import { ChevronRight } from '@material-ui/icons'
-
-const useStyles = makeStyles({
-  tituloContainer: {
-    display: 'flex',
-    alignItems: 'flex-end',
-    marginBottom: 48
-  },
-  titulo: {
-    flex: 1,
-    textAlign: 'center'
-  }
-})
+import TituloConfig from 'src/components/TituloConfig'
 
 const MeusDados = () => {
-  const classes = useStyles()
   const router = useRouter()
 
   const handleAlterarNome = () => {
@@ -40,12 +27,7 @@ const MeusDados = () => {
 
   return (
     <Container maxWidth="xs">
-      <Box className={classes.tituloContainer}>
-        <LinkVoltar href="/app/perfil" />
-        <Typography variant="button" className={classes.titulo}>
-          Meus dados
-        </Typography>
-      </Box>
+      <TituloConfig link="/app/perfil/dados" titulo="Meus dados" />
       <NavigationList itens={itens} />
     </Container>
   )
